Update README.md
The Emoji Catcher Game is a beginner-friendly browser game built using HTML, CSS, and  JavaScript. The player has 60 seconds to catch as many falling emojis as possible by clicking on them before they disappear. It features a simple UI, real-time score tracking, and a countdown timer.

This project is ideal for those who are new to front-end development and want to practice DOM manipulation, event handling, and basic game logic. It’s also a fun way to understand how to build interactive elements and timed behaviors in a web application.